+++ lib/src/solid/utils/misc.dart
@@ -30,42 +30,15 @@
 
 library;
 
-import 'dart:convert' show utf8;
-
-import 'package:flutter/foundation.dart' show debugPrint;
-
-import 'package:encrypter_plus/encrypter_plus.dart';
-import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
-import 'package:jwt_decoder/jwt_decoder.dart';
-import 'package:path/path.dart' as path;
-import 'package:rdflib/rdflib.dart';
-import 'package:solid_auth/solid_auth.dart' show DpopTokenGenerator;
 
 import 'package:solidpod/src/solid/api/rest_api.dart';
 import 'package:solidpod/src/solid/constants/common.dart';
-import 'package:solidpod/src/solid/constants/path_type.dart';
-import 'package:solidpod/src/solid/constants/schema.dart';
-import 'package:solidpod/src/solid/public_sharing_hooks.dart';
-import 'package:solidpod/src/solid/utils/app_info.dart';
-import 'package:solidpod/src/solid/utils/authdata_manager.dart';
-import 'package:solidpod/src/solid/utils/data_encryption.dart';
 import 'package:solidpod/src/solid/utils/get_url_helper.dart';
-import 'package:solidpod/src/solid/utils/key_helper.dart' show getPredicateUrl;
-import 'package:solidpod/src/solid/utils/key_manager.dart';
-import 'package:solidpod/src/solid/utils/rdf.dart';
 
-/// Global callback for clearing application-specific caches during logout.
-/// Apps should register their cache clearing logic here.
-/// This ensures caches are cleared BEFORE any blocking network operations.
-Future<void> Function()? _onLogoutClearCaches;
-
-/// Register a callback to clear application-specific caches during logout.
-/// This callback will be invoked BEFORE the OAuth2 logout endpoint call,
-/// preventing race conditions where cached data might be visible during logout.
-void registerLogoutCacheCallback(Future<void> Function() callback) {
-  _onLogoutClearCaches = callback;
-}
+export 'package:solidpod/src/solid/utils/enc_in_place.dart';
+export 'package:solidpod/src/solid/utils/pod_paths.dart';
+export 'package:solidpod/src/solid/utils/session.dart';
 
 // solid-encrypt uses unencrypted local storage and refers to http: //yarrabah.net/ for predicates definition,
 // do not use it before it is updated (same as what the gurriny project does)
@@ -87,447 +60,6 @@ Future<void> writeToSecureStorage(String key, String value) async {
   }
 
   await secureStorage.write(key: key, value: value);
-}
-
-/// Load and parse a private TTL file from POD
-// Future<Map<String, dynamic>> loadPrvTTL(String fileUrl) async {
-//   // final fileUrl = await getFileUrl(filePath);
-//   try {
-//     if (await checkResourceStatus(fileUrl) == ResourceStatus.exist) {
-//       final rawContent = await fetchPrvFile(fileUrl);
-//       return parseTTL(rawContent);
-//     } else {
-//       return {};
-//     }
-//   } on Exception catch (e) {
-//     throw Exception(e);
-//   }
-// }
-
-/// Read the encryption key file content for display purposes.
-///
-/// This function directly reads the encryption key file without using readPod,
-/// making it suitable for accessing files outside the appname/data directory.
-///
-/// Returns the raw TTL content of the encryption key file.
-// Future<String> readEncryptionKeyContent() async {
-//   final encKeyPath = await getEncKeyPath();
-//   final encKeyUrl = await getFileUrl(encKeyPath);
-
-//   try {
-//     if (await checkResourceStatus(encKeyUrl) == ResourceStatus.exist) {
-//       return utf8.decode(
-//         await getResource(encKeyUrl),
-//       );
-
-//       return await fetchPrvFile(encKeyUrl);
-//     } else {
-//       throw Exception('Encryption key file does not exist at: $encKeyPath');
-//     }
-//   } on Exception catch (e) {
-//     throw Exception('Failed to read encryption key file: $e');
-//   }
-// }
-
-/// Encrypt a given data string and format to TTL
-Future<String> getEncTTLStr({
-  required String fileUrl,
-  required String fileContent,
-  required Key key,
-  required IV iv,
-  String? inheritKeyFrom,
-}) async {
-  final filePath = await extractResourcePathFromUrl(fileUrl);
-  final triples = {
-    URIRef(fileUrl): {
-      solidTermsNS.ns.withAttr(pathPred): filePath,
-      solidTermsNS.ns.withAttr(ivPred): iv.base64,
-      if (inheritKeyFrom != null)
-        solidTermsNS.ns.withAttr(inheritKeyPred): inheritKeyFrom,
-      solidTermsNS.ns.withAttr(encDataPred): encryptData(fileContent, key, iv),
-    },
-  };
-
-  final bindNS = {solidTermsNS.prefix: solidTermsNS.ns};
-
-  return tripleMapToTurtle(triples, bindNamespaces: bindNS);
-}
-
-/// Extract the `(iv, encData)` pair from the encrypted-TTL wrapper written
-/// by [getEncTTLStr] for [fileUrl], or return `null` if [rawTtl] is not a
-/// solidpod-encrypted document. Used by the in-place sharing-lifecycle
-/// helpers below to recover the ciphertext alongside its IV in one pass.
-
-({String iv, String encData})? _extractEncFields(
-  String rawTtl,
-  String fileUrl,
-) {
-  Map<String, Map<String, dynamic>> tripleMap;
-  try {
-    tripleMap = turtleToTripleMap(rawTtl);
-  } on Object catch (e) {
-    debugPrint('[_extractEncFields] failed to parse "$fileUrl": $e');
-    return null;
-  }
-
-  final ivKey = getPredicateUrl(ivPred);
-  final encKey = getPredicateUrl(encDataPred);
-
-  ({String iv, String encData})? pickFrom(Map<String, dynamic> m) {
-    final iv = m[ivKey];
-    final enc = m[encKey];
-    if (iv is String && enc is String) {
-      return (iv: iv, encData: enc);
-    }
-    return null;
-  }
-
-  final direct = tripleMap[fileUrl];
-  if (direct != null) {
-    final r = pickFrom(direct);
-    if (r != null) return r;
-  }
-
-  // Fall back to any subject in the document. This covers ttl that was
-  // written under a slightly different subject URL (e.g. due to a host
-  // rewrite between writePod and the current read).
-
-  for (final entry in tripleMap.entries) {
-    if (entry.key == fileUrl) continue;
-    final r = pickFrom(entry.value);
-    if (r != null) {
-      debugPrint(
-        '[_extractEncFields] iv/encData found under fallback subject '
-        '"${entry.key}" instead of "$fileUrl"',
-      );
-      return r;
-    }
-  }
-
-  return null;
-}
-
-/// Decrypt the content of [fileUrl] in place on the server, replacing the
-/// encrypted TTL payload with the plaintext that was originally written.
-///
-/// Intended for use by [grantPermission] when sharing a file with the
-/// Public or Authenticated User class: those recipients cannot be issued
-/// a per-recipient encryption key, so the only way for them to actually
-/// read the resource by navigating to its URL is for the file itself to
-/// be plaintext on the server. The individual key is intentionally kept
-/// in `ind-keys.ttl` so that the file can be re-encrypted later by
-/// [encryptFileInPlace] if public/auth-user access is revoked.
-///
-/// After unwrapping solidpod's own encrypted-TTL layer, the
-/// [PublicSharingHooks.onPublicShareDecrypted] hook (if registered) is
-/// applied so that a host app can also strip an application-level
-/// encryption layer (e.g. NotePod's per-note noteContent ciphertext)
-/// before the bytes are exposed to anonymous readers.
-///
-/// No-op when the file is not currently in the encrypted TTL format.
-/// Throws when the individual encryption key cannot be obtained — the
-/// caller is responsible for deciding whether to surface that to the
-/// user or skip the publish step.
-
-Future<void> decryptFileInPlace(
-  String fileUrl, {
-  bool isExternalRes = false,
-}) async {
-  debugPrint('[decryptFileInPlace] start url="$fileUrl" '
-      'isExternalRes=$isExternalRes');
-  final raw = utf8.decode(await getResource(fileUrl));
-  final encMap = _extractEncFields(raw, fileUrl);
-
-  if (encMap == null) {
-    debugPrint(
-      '[decryptFileInPlace] file is not in encrypted format, nothing to do: '
-      '"$fileUrl"',
-    );
-    return;
-  }
-
-  final indKey = isExternalRes
-      ? await KeyManager.getSharedIndividualKey(fileUrl)
-      : await KeyManager.getIndividualKey(fileUrl);
-
-  if (indKey == null) {
-    throw Exception(
-      'No individual encryption key available for "$fileUrl"; '
-      'cannot decrypt the file for public/authenticated-user sharing.',
-    );
-  }
-
-  var plaintext = decryptData(
-    encMap.encData,
-    indKey,
-    IV.fromBase64(encMap.iv),
-  );
-
-  // Allow the host app to strip an additional application-level
-  // encryption layer before the file is exposed publicly.
-
-  final postHook = PublicSharingHooks.onPublicShareDecrypted;
-  if (postHook != null) {
-    debugPrint(
-      '[decryptFileInPlace] applying app onPublicShareDecrypted hook',
-    );
-    plaintext = await postHook(fileUrl, plaintext);
-  }
-
-  await createResource(
-    fileUrl,
-    content: plaintext,
-    contentType: ResourceContentType.turtleText,
-  );
-  debugPrint('[decryptFileInPlace] wrote plaintext (${plaintext.length} '
-      'bytes) to "$fileUrl"');
-}
-
-/// Apply the [PublicSharingHooks.onPublicShareDecrypted] transformer to
-/// the current (already-plaintext) bytes of [fileUrl] and write back any
-/// resulting change.
-///
-/// Used by [grantPermission] when the outer encrypted-TTL wrapper has
-/// already been removed (e.g. by an older solidpod that decrypted the
-/// file but didn't run the application-level hook). Without this, a
-/// file that was left in such a mixed state would never have its inner
-/// ciphertext stripped.
-
-Future<void> applyPublicShareDecryptedHookInPlace(String fileUrl) async {
-  final hook = PublicSharingHooks.onPublicShareDecrypted;
-  if (hook == null) {
-    debugPrint(
-      '[applyPublicShareDecryptedHookInPlace] no hook registered, '
-      'nothing to do: "$fileUrl"',
-    );
-    return;
-  }
-
-  try {
-    final current = utf8.decode(await getResource(fileUrl));
-    final transformed = await hook(fileUrl, current);
-    if (transformed == current) {
-      debugPrint(
-        '[applyPublicShareDecryptedHookInPlace] hook left content '
-        'unchanged: "$fileUrl"',
-      );
-      return;
-    }
-    await createResource(
-      fileUrl,
-      content: transformed,
-      contentType: ResourceContentType.turtleText,
-    );
-    debugPrint(
-      '[applyPublicShareDecryptedHookInPlace] hook rewrote content '
-      '(${transformed.length} bytes) for "$fileUrl"',
-    );
-  } on Object catch (e) {
-    debugPrint(
-      '[applyPublicShareDecryptedHookInPlace] hook failed for "$fileUrl": $e',
-    );
-  }
-}
-
-/// Encrypt the (currently plaintext) content of [fileUrl] in place using
-/// the individual encryption key already recorded for the resource.
-///
-/// The counterpart of [decryptFileInPlace], called by [revokePermission]
-/// when public/auth-user access is removed so that the resource's
-/// at-rest representation is restored to what the host app originally
-/// wrote via `writePod`. No-op when the file is already in the
-/// encrypted TTL format or when no individual key is available
-/// (i.e. the file was always plaintext at rest).
-///
-/// The [PublicSharingHooks.onPublicShareRevoked] hook (if registered)
-/// is applied before re-wrapping so that the host app can restore the
-/// application-level encryption layer that [decryptFileInPlace] peeled
-/// off when sharing was granted.
-
-Future<void> encryptFileInPlace(
-  String fileUrl, {
-  bool isExternalRes = false,
-}) async {
-  debugPrint('[encryptFileInPlace] start url="$fileUrl" '
-      'isExternalRes=$isExternalRes');
-  final indKey = isExternalRes
-      ? await KeyManager.getSharedIndividualKey(fileUrl)
-      : await KeyManager.getIndividualKey(fileUrl);
-  if (indKey == null) {
-    debugPrint('[encryptFileInPlace] no individual key for "$fileUrl", '
-        'leaving file as plaintext');
-    return;
-  }
-
-  final raw = utf8.decode(await getResource(fileUrl));
-  if (_extractEncFields(raw, fileUrl) != null) {
-    debugPrint('[encryptFileInPlace] file already encrypted, nothing to do: '
-        '"$fileUrl"');
-    return;
-  }
-
-  var plaintext = raw;
-
-  // Give the host app a chance to restore an inner application-level
-  // encryption layer that was peeled off by [decryptFileInPlace] when
-  // public/auth-user sharing was granted, so that at-rest state is
-  // symmetric with what the app originally wrote via writePod.
-
-  final preHook = PublicSharingHooks.onPublicShareRevoked;
-  if (preHook != null) {
-    debugPrint(
-      '[encryptFileInPlace] applying app onPublicShareRevoked hook',
-    );
-    plaintext = await preHook(fileUrl, plaintext);
-  }
-
-  final encContent = await getEncTTLStr(
-    fileUrl: fileUrl,
-    fileContent: plaintext,
-    key: indKey,
-    iv: IV.fromLength(16),
-  );
-
-  await createResource(
-    fileUrl,
-    content: encContent,
-    contentType: ResourceContentType.turtleText,
-  );
-  debugPrint('[encryptFileInPlace] re-encrypted "$fileUrl"');
-}
-
-/// Returns `true` when [fileUrl] currently holds an encrypted-TTL payload
-/// written by [getEncTTLStr]. Cheaper than [getFileEncryptionStatus] when
-/// the caller only needs the boolean and is happy to silently treat any
-/// fetch / parse failure as "not encrypted".
-
-Future<bool> isFileContentEncrypted(String fileUrl) async {
-  if (!fileUrl.toLowerCase().endsWith('.ttl')) {
-    return false;
-  }
-  try {
-    if (await checkResourceStatus(fileUrl) != ResourceStatus.exist) {
-      return false;
-    }
-    final raw = utf8.decode(await getResource(fileUrl));
-    return _extractEncFields(raw, fileUrl) != null;
-  } on Object catch (e) {
-    debugPrint('[isFileContentEncrypted] unable to inspect "$fileUrl": $e');
-    return false;
-  }
-}
-
-/// Returns the path of file with verification key and private key
-Future<String> getEncKeyPath() async =>
-    [appDirName, encDir, encKeyFile].join('/');
-
-/// Returns the path of file with individual keys
-Future<String> getIndKeyPath() async =>
-    [appDirName, encDir, indKeyFile].join('/');
-
-/// Returns the path of file with public keys
-Future<String> getPubKeyPath() async =>
-    [appDirName, sharingDir, pubKeyFile].join('/');
-
-/// Returns the path of public file with individual keys
-Future<String> getPubIndKeyPath() async =>
-    [appDirName, sharingDir, pubIndKeyFile].join('/');
-
-/// Returns the path of file with individual keys accessed only
-/// by authenticated users
-Future<String> getAuthUserIndKeyPath() async =>
-    [appDirName, sharingDir, authUserIndKeyFile].join('/');
-
-/// Returns the path of the data directory
-Future<String> getDataDirPath() async => [appDirName, dataDir].join('/');
-
-/// Checks whether a POD-relative [resourcePath] falls within the current
-/// application's directory tree.
-///
-/// Returns `true` if the resource belongs to this app, meaning the app
-/// holds the encryption key required to decrypt it. Returns `false` if
-/// the resource belongs to another application's folder, in which case
-/// decryption may not be possible.
-///
-/// [resourcePath] should be a normalised POD-relative path (e.g.
-/// `myapp/data/file.ttl`). Absolute URLs or empty strings return `false`.
-
-Future<bool> isPathInCurrentApp(String resourcePath) async {
-  try {
-    if (resourcePath.trim().isEmpty) return false;
-
-    if (resourcePath.startsWith('http://') ||
-        resourcePath.startsWith('https://')) {
-      debugPrint(
-        'isPathInCurrentApp: expected a POD-relative path but received '
-        'an absolute URL: $resourcePath',
-      );
-      return false;
-    }
-
-    // Derive the current app name from getDataDirPath() which returns
-    // "APP_NAME/data". The first segment is the app name.
-
-    final appDataPath = await getDataDirPath();
-    if (appDataPath.isEmpty) return false;
-
-    final currentAppName = appDataPath.split('/').first;
-    if (currentAppName.isEmpty) return false;
-
-    // Build full URLs for both the resource and the app root, then
-    // compare prefixes. getDirUrl appends a trailing slash which
-    // prevents false positives (e.g. "myapp2" matching "myapp").
-
-    final resourceUrl = await getFileUrl(resourcePath);
-    final appRootUrl = await getDirUrl(currentAppName);
-
-    return resourceUrl.startsWith(appRootUrl);
-  } catch (e) {
-    debugPrint('Error in isPathInCurrentApp: $e');
-    return false;
-  }
-}
-
-/// Returns the path of the shared directory
-Future<String> getSharedDirPath() async => [appDirName, sharedDir].join('/');
-
-/// Returns the path of the file with shared individual keys
-Future<String> getSharedKeyFilePath() async =>
-    [appDirName, sharedDir, sharedKeyFile].join('/');
-
-/// Returns the path of the encryption directory
-Future<String> getEncDirPath() async => [appDirName, encDir].join('/');
-
-/// Returns the path of the encryption directory
-Future<String> getPermLogFilePath() async =>
-    [appDirName, logsDir, permLogFile].join('/');
-
-/// Extract the app name and the version from the package info
-/// Return a record (with named fields https://dart.dev/language/records)
-
-Future<({String name, String version})> getAppNameVersion() async =>
-    (name: await AppInfo.name, version: await AppInfo.version);
-
-/// Return the web ID
-Future<String?> getWebId() async => AuthDataManager.getWebId();
-
-/// Check whether a user is logged in or not
-///
-/// Check if the local storage has authentication
-/// details of the user and also check whether the
-/// access token is expired or not
-
-Future<bool> isUserLoggedIn() async {
-  final webId = await AuthDataManager.getWebId();
-
-  if (webId != null && webId.isNotEmpty) {
-    final accessToken = await AuthDataManager.getAccessToken();
-    if (accessToken != null && !JwtDecoder.isExpired(accessToken)) {
-      return true;
-    }
-  }
-
-  return false;
 }
 
 /// Create a directory with the given URL.
@@ -604,35 +136,6 @@ Future<void> createContainer(String parentPath, String folderName) async {
   await createDir(dirUrl);
 }
 
-/// Delete login information from the local storage
-///
-/// returns true if successful
-
-Future<bool> deleteLogIn() async => AuthDataManager.removeAuthData();
-
-/// Set directory name for the app for storing the POD data
-///
-/// If not initially set the app name will be taken by default.
-
-Future<void> setAppDirName(String inputAppDirName) async {
-  if (inputAppDirName.isEmpty) {
-    appDirName = await AppInfo.canonicalName;
-  } else {
-    appDirName = inputAppDirName;
-  }
-}
-
-/// Get resource acl file path
-String getResAclFile(String resourceUrl, [bool isFile = true]) {
-  final resourceAclUrl = resourceUrl.endsWith('.acl')
-      ? resourceUrl
-      : isFile
-          ? '$resourceUrl.acl'
-          : '$resourceUrl/.acl';
-
-  return resourceAclUrl;
-}
-
 /// Extract permission details of a file into a map.
 /// Returns a map where keys are permission receiver webIds and
 /// values are the list of permissions
@@ -682,129 +185,6 @@ String getResNameFromUrl(String resourceUrl) {
   return resourceUrl.split('/').last;
 }
 
-/// Get tokens necessary to fetch a resource from a POD
-///
-/// returns the access token and DPoP token
-Future<({String accessToken, String dPopToken})> getTokensForResource(
-  String resourceUrl,
-  String httpMethod,
-) async {
-  final authData = await AuthDataManager.loadAuthData();
-
-  if (authData == null) {
-    throw Exception('Authentication data not available. Please login first.');
-  }
-
-  final authManager = AuthDataManager.getAuthManager();
-  if (authManager == null) {
-    throw Exception('Auth manager not available. Please login first.');
-  }
-
-  final dPopToken = await DpopTokenGenerator.generateForRequest(
-    endpointUrl: resourceUrl,
-    httpMethod: httpMethod,
-    accessToken: authData.accessToken,
-    keyManager: authManager.keyManager,
-  );
-
-  return (
-    accessToken: authData.accessToken,
-    dPopToken: dPopToken,
-  );
-}
-
-/// Logging out the user with comprehensive error handling and platform support
-///
-/// This function performs a complete logout that includes:
-/// 1. Clearing all encryption keys from memory
-/// 2. Clearing application-specific caches
-/// 3. Calling the OIDC logout endpoint via SolidAuthManager (with error tolerance)
-/// 4. Removing authentication data from secure storage
-///
-/// Returns true if critical cleanup (key/auth data) succeeds.
-Future<bool> logoutPod() async {
-  try {
-    await KeyManager.clear();
-
-    // Clear app caches before any network operations to prevent race conditions.
-    if (_onLogoutClearCaches != null) {
-      try {
-        await _onLogoutClearCaches!();
-      } on Object catch (e) {
-        debugPrint('logoutPod() cache callback failed (non-critical): $e');
-      }
-    }
-
-    // Contact the OIDC logout endpoint and rotate the DPoP key.
-    // Must be called before removeAuthData() clears _authManager.
-    try {
-      await AuthDataManager.getAuthManager()?.logout();
-    } on Object catch (e) {
-      debugPrint('logoutPod() OAuth2 logout warning (non-critical): $e');
-    }
-
-    final authDataRemoved = await AuthDataManager.removeAuthData();
-    return authDataRemoved;
-  } on Object catch (e) {
-    debugPrint('logoutPod() CRITICAL ERROR: $e');
-    try {
-      await AuthDataManager.removeAuthData();
-      await KeyManager.clear();
-    } on Object catch (_) {}
-    return false;
-  }
-}
-
-/// Clear all login state without opening a browser.
-///
-/// Performs the same cleanup as [logoutPod] but invalidates the IdP session
-/// via a headless HTTP request rather than launching a visible browser
-/// window. Use this when switching accounts or recovering from stale
-/// credentials.
-
-Future<bool> silentLogout() async {
-  try {
-    await KeyManager.clear();
-
-    // Get logout URL from discovery doc before clearing the manager.
-    String? logoutUrl;
-    try {
-      logoutUrl = await AuthDataManager.getLogoutUrl();
-    } on Object catch (_) {}
-
-    // Clear local token state only — no browser redirect.
-    await AuthDataManager.getAuthManager()?.forgetUser();
-
-    final authDataRemoved = await AuthDataManager.removeAuthData();
-
-    if (_onLogoutClearCaches != null) {
-      try {
-        await _onLogoutClearCaches!();
-      } on Object catch (e) {
-        debugPrint('silentLogout() cache callback failed (non-critical): $e');
-      }
-    }
-
-    // Best-effort IdP session invalidation via headless HTTP GET.
-    if (logoutUrl != null && logoutUrl.isNotEmpty) {
-      try {
-        await http.get(Uri.parse(logoutUrl));
-      } on Object catch (e) {
-        debugPrint('silentLogout() headless logout failed (non-critical): $e');
-      }
-    }
-
-    return authDataRemoved;
-  } on Object catch (e) {
-    debugPrint('silentLogout() CRITICAL: $e');
-    try {
-      await AuthDataManager.removeAuthData();
-      await KeyManager.clear();
-    } on Object catch (_) {}
-    return false;
-  }
-}
-
 /// Removes header and footer (which mess up the TTL format) from a PEM-formatted public key string.
 ///
 /// This function takes a public key string, typically in PEM format, and removes
@@ -828,108 +208,4 @@ String getDateTime(String dateTimeStr) {
   final dateFormat = DateFormat('dd/MM/yyyy hh:mm:ss a');
 
   return dateFormat.format(dateTime);
-}
-
-/// Normalise file path for readPod/writePod operations.
-///
-/// Handles backward compatibility by checking if the filePath already includes
-/// the app directory prefix, and constructs the appropriate normalised path.
-///
-/// When basePath is null (default for readPod/writePod), uses appname/data as base path.
-///
-/// [filePath] - The input file path
-/// [basePath] - The base path to use (defaults to appname/data when null)
-///
-/// Returns the normalised file path.
-///
-/// Examples:
-/// - `normalizeFilePath('abc.ttl', null)` returns `appname/data/abc.ttl`
-/// - `normalizeFilePath('movies/abc.ttl', null)` returns `appname/data/movies/abc.ttl`
-/// - `normalizeFilePath('appname/data/keys.ttl', null)` returns `appname/data/keys.ttl`
-///
-/// Note: Only `appname/data/` paths are supported for readPod/writePod operations.
-
-Future<String> normalizeFilePath(String filePath, String? basePath) async {
-  // Normalise path separators for cross-platform compatibility.
-
-  final normalizedInput = filePath.replaceAll(path.separator, '/');
-
-  // Use provided path or default to appname/data.
-
-  final effectiveBasePath = basePath == null || basePath.trim().isEmpty
-      ? await getDataDirPath()
-      : basePath;
-
-  // Check if path already starts with the correct base path (appname/data/).
-
-  if (normalizedInput.startsWith(effectiveBasePath)) {
-    // Full path is already prepended (appname/data/).
-
-    return normalizedInput;
-  } else {
-    // Prepend the base path.
-
-    return [effectiveBasePath, normalizedInput].join('/');
-  }
-}
-
-/// Check if a given path string is a directory or not
-bool isDir(String path) {
-  if (path.endsWith('/') || !path.contains('.')) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-/// Generate the URL of resource according to its path and the type of the path.
-
-Future<String> generateResourceUrlFromPath({
-  required String resourcePath,
-  required PathType pathType,
-  bool isFile = true,
-  String? webId,
-}) async {
-  final func = isFile ? getFileUrl : getDirUrl;
-  switch (pathType) {
-    case PathType.absoluteUrl:
-      return resourcePath;
-
-    case PathType.relativeToPod:
-      return await func(resourcePath, webId: webId);
-
-    case PathType.relativeToApp:
-      return await func([appDirName, resourcePath].join('/'), webId: webId);
-
-    case PathType.relativeToData:
-      return await func(
-        [await getDataDirPath(), resourcePath].join('/'),
-        webId: webId,
-      );
-  }
-}
-
-/// Extract resource path from its URL
-/// path format:
-/// - appDir/path/to/file
-/// - appDir/path/to/dir/
-
-Future<String> extractResourcePathFromUrl(
-  String resourceUrl, {
-  bool isFile = true,
-}) async {
-  // See https://api.dart.dev/dart-core/Uri-class.html for details
-
-  final segments = Uri.parse(resourceUrl).pathSegments;
-
-  final path = segments.getRange(1, segments.length).join('/');
-
-  return !(isFile || path.endsWith('/')) ? '$path/' : path;
-}
-
-/// Generate the Web ID of from resource URL
-
-Future<String> generateWebIdFromResourceUrl(String resourceUrl) async {
-  final uri = Uri.parse(resourceUrl);
-  return [uri.origin, uri.pathSegments.first, profCard].join('/');
 }
